requirements.txt
certifi==2019.9.11
chardet==3.0.4
cycler==0.10.0
idna==2.8
imageio==2.6.1
kiwisolver==1.1.0
matplotlib==3.1.1
numpy==1.17.3
Pillow==6.2.1
pyparsing==2.4.4
pyTelegramBotAPI==3.6.6
python-dateutil==2.8.1
requests==2.22.0
scipy==1.3.2
six==1.12.0
urllib3==1.25.6
